requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
jinja2==3.1.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
